feat(index.tsx): moved / -> /login and /roleSelection -> /
